Refractor class names
